+++ data/remote/src/main/kotlin/com/berlin/remote/TvShowDetailsRemoteDataSourceImpl.kt
@@ -10,7 +10,7 @@ import io.ktor.client.request.parameter
 class TvShowDetailsRemoteDataSourceImpl(
     private val client: HttpClient
 ) : TvShowDetailsRemoteDataSource {
-    override suspend fun getTvShowDetails(id: Int, language: String): TVShowDetailsDto {
+    override suspend fun getTvShowDetails(id: Long, language: String): TVShowDetailsDto {
         return client.get("tv/$id"){
             parameter(ApiConstants.LANGUAGE, language)
         }.body()
