+++ data/repository/src/main/java/com/berlin/repository/datasource/remote/MovieDetailsRemoteDataSource.kt
@@ -3,5 +3,5 @@ package com.berlin.repository.datasource.remote
 import com.berlin.repository.datasource.remote.dto.MovieDetailsDto
 
 interface MovieDetailsRemoteDataSource {
-    suspend fun getMovieDetails(id: Int,language: String): MovieDetailsDto
+    suspend fun getMovieDetails(id: Long,language: String): MovieDetailsDto
 }
