+++ data/repository/src/main/java/com/berlin/repository/datasource/remote/TvShowDetailsRemoteDataSource.kt
@@ -3,5 +3,5 @@ package com.berlin.repository.datasource.remote
 import com.berlin.repository.datasource.remote.dto.TVShowDetailsDto
 
 interface TvShowDetailsRemoteDataSource {
-    suspend fun getTvShowDetails(id: Int,language: String): TVShowDetailsDto
+    suspend fun getTvShowDetails(id: Long,language: String): TVShowDetailsDto
 }
